+++ MinistryofJustice/src/mj/access/grp/AP_REPORT_TYPE.java
@@ -1,22 +1,22 @@
 package mj.access.grp;
 
-import javafx.beans.property.IntegerProperty;
-import javafx.beans.property.SimpleIntegerProperty;
+import javafx.beans.property.LongProperty;
+import javafx.beans.property.SimpleLongProperty;
 import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
 
 public class AP_REPORT_TYPE {
 	/** Идентификатор типа отчета */
-	private IntegerProperty REPORT_TYPE_ID;
+	private LongProperty REPORT_TYPE_ID;
 	/** Наименование типа отчета */
 	private StringProperty REPORT_TYPE_NAME;
 
 	public AP_REPORT_TYPE() {
-		this.REPORT_TYPE_ID = new SimpleIntegerProperty();
+		this.REPORT_TYPE_ID = new SimpleLongProperty();
 		this.REPORT_TYPE_NAME = new SimpleStringProperty();
 	}
 
-	public void setREPORT_TYPE_ID(Integer REPORT_TYPE_ID) {
+	public void setREPORT_TYPE_ID(Long REPORT_TYPE_ID) {
 		this.REPORT_TYPE_ID.set(REPORT_TYPE_ID);
 	}
 
@@ -24,7 +24,7 @@ public class AP_REPORT_TYPE {
 		this.REPORT_TYPE_NAME.set(REPORT_TYPE_NAME);
 	}
 
-	public Integer getREPORT_TYPE_ID() {
+	public Long getREPORT_TYPE_ID() {
 		return REPORT_TYPE_ID.get();
 	}
 
@@ -32,7 +32,7 @@ public class AP_REPORT_TYPE {
 		return REPORT_TYPE_NAME.get();
 	}
 
-	public IntegerProperty REPORT_TYPE_IDProperty() {
+	public LongProperty REPORT_TYPE_IDProperty() {
 		return REPORT_TYPE_ID;
 	}
