+++ MinistryofJustice/src/mj/access/grp/ODB_GROUP_USR.java
@@ -1,25 +1,25 @@
 package mj.access.grp;
 
-import javafx.beans.property.IntegerProperty;
-import javafx.beans.property.SimpleIntegerProperty;
+import javafx.beans.property.LongProperty;
+import javafx.beans.property.SimpleLongProperty;
 import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
 
 public class ODB_GROUP_USR {
 	/** Идентификатор группы */
-	private IntegerProperty GRP_ID;
+	private LongProperty GRP_ID;
 	/** Наименование группы */
 	private StringProperty GRP_NAME;
 	/** Название */
 	private StringProperty NAME;
 
 	public ODB_GROUP_USR() {
-		this.GRP_ID = new SimpleIntegerProperty();
+		this.GRP_ID = new SimpleLongProperty();
 		this.GRP_NAME = new SimpleStringProperty();
 		this.NAME = new SimpleStringProperty();
 	}
 
-	public void setGRP_ID(Integer GRP_ID) {
+	public void setGRP_ID(Long GRP_ID) {
 		this.GRP_ID.set(GRP_ID);
 	}
 
@@ -31,7 +31,7 @@ public class ODB_GROUP_USR {
 		this.NAME.set(NAME);
 	}
 
-	public Integer getGRP_ID() {
+	public Long getGRP_ID() {
 		return GRP_ID.get();
 	}
 
@@ -43,7 +43,7 @@ public class ODB_GROUP_USR {
 		return NAME.get();
 	}
 
-	public IntegerProperty GRP_IDProperty() {
+	public LongProperty GRP_IDProperty() {
 		return GRP_ID;
 	}
