+++ MinistryofJustice/src/mj/access/menu/ODB_MNU.java
@@ -1,36 +1,36 @@
 package mj.access.menu;
 
-import javafx.beans.property.IntegerProperty;
-import javafx.beans.property.SimpleIntegerProperty;
+import javafx.beans.property.LongProperty;
+import javafx.beans.property.SimpleLongProperty;
 import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
 
 public class ODB_MNU {
 	/** Идентификатор действия */
-	private IntegerProperty MNU_ID;
+	private LongProperty MNU_ID;
 	/** Идентификатор родительского действия */
-	private IntegerProperty MNU_PARENT;
+	private LongProperty MNU_PARENT;
 	/** Порядковый номер внутри иерархии */
-	private IntegerProperty MNU_NPP;
+	private LongProperty MNU_NPP;
 	/** Наименование действия */
 	private StringProperty MNU_NAME;
 
 	public ODB_MNU() {
-		this.MNU_ID = new SimpleIntegerProperty();
-		this.MNU_PARENT = new SimpleIntegerProperty();
-		this.MNU_NPP = new SimpleIntegerProperty();
+		this.MNU_ID = new SimpleLongProperty();
+		this.MNU_PARENT = new SimpleLongProperty();
+		this.MNU_NPP = new SimpleLongProperty();
 		this.MNU_NAME = new SimpleStringProperty();
 	}
 
-	public void setMNU_ID(Integer MNU_ID) {
+	public void setMNU_ID(Long MNU_ID) {
 		this.MNU_ID.set(MNU_ID);
 	}
 
-	public void setMNU_PARENT(Integer MNU_PARENT) {
+	public void setMNU_PARENT(Long MNU_PARENT) {
 		this.MNU_PARENT.set(MNU_PARENT);
 	}
 
-	public void setMNU_NPP(Integer MNU_NPP) {
+	public void setMNU_NPP(Long MNU_NPP) {
 		this.MNU_NPP.set(MNU_NPP);
 	}
 
@@ -38,15 +38,15 @@ public class ODB_MNU {
 		this.MNU_NAME.set(MNU_NAME);
 	}
 
-	public Integer getMNU_ID() {
+	public Long getMNU_ID() {
 		return MNU_ID.get();
 	}
 
-	public Integer getMNU_PARENT() {
+	public Long getMNU_PARENT() {
 		return MNU_PARENT.get();
 	}
 
-	public Integer getMNU_NPP() {
+	public Long getMNU_NPP() {
 		return MNU_NPP.get();
 	}
 
@@ -54,15 +54,15 @@ public class ODB_MNU {
 		return MNU_NAME.get();
 	}
 
-	public IntegerProperty MNU_IDProperty() {
+	public LongProperty MNU_IDProperty() {
 		return MNU_ID;
 	}
 
-	public IntegerProperty MNU_PARENTProperty() {
+	public LongProperty MNU_PARENTProperty() {
 		return MNU_PARENT;
 	}
 
-	public IntegerProperty MNU_NPPProperty() {
+	public LongProperty MNU_NPPProperty() {
 		return MNU_NPP;
 	}
